[WNMGDS-795] Reorganize and rename form components (#929)
* Rename FieldContainer files to FormControl

* Replace instances of FieldContainer with FormControl

* Add todo comments for improving FormControl

* Moved InlineError into separate component
