+++ packages/design-system/src/components/FormControl/FormControl.tsx
@@ -1,18 +1,18 @@
 import FormLabel from '../FormLabel/FormLabel';
-import InlineError from './InlineError';
+import InlineError from '../InlineError/InlineError';
 import React from 'react';
 import classNames from 'classnames';
 import { errorPlacementDefault } from '../flags';
 import { uniqueId } from 'lodash';
 
-interface FieldContainerRenderProps {
+interface FormControlRenderProps {
   id: string;
   labelId: string;
   errorId: string;
   setRef: (elem: HTMLDivElement) => void;
 }
 
-interface FieldContainerProps {
+interface FormControlProps {
   /**
    * Additional classes to be added to the field container.
    */
@@ -77,11 +77,11 @@ interface FieldContainerProps {
   /**
    * A function that returns a field input element to accept render props
    */
-  render: (renderProps: FieldContainerRenderProps) => React.ReactNode;
+  render: (renderProps: FormControlRenderProps) => React.ReactNode;
 }
 
-export class FieldContainer extends React.Component<FieldContainerProps> {
-  constructor(props: FieldContainerProps) {
+export class FormControl extends React.Component<FormControlProps> {
+  constructor(props: FormControlProps) {
     super(props);
 
     this.id = props.id || uniqueId('field_');
@@ -93,7 +93,6 @@ export class FieldContainer extends React.Component<FieldContainerProps> {
   componentDidMount(): void {	
     // Automatically set focus on field input element when `focusTrigger` prop is used
     if (this.props.focusTrigger && this.inputRef) {	
-      console.log(this.inputRef)
       this.inputRef.focus();	
     }	
   }
@@ -144,7 +143,7 @@ export class FieldContainer extends React.Component<FieldContainerProps> {
     // Errored form components without fieldsets must handle `aria-invalid` in their own component
     const ariaInvalid = isFieldset && errorMessage ? true : undefined;
 
-    // Bottom placed errors are handled in FieldContainer instead of FormLabel
+    // Bottom placed errors are handled in FormControl instead of FormLabel
     const renderBottomError = bottomError ? (
       <InlineError id={this.errorId} inversed={inversed} className={errorMessageClassName}>
         {errorMessage}
@@ -157,7 +156,10 @@ export class FieldContainer extends React.Component<FieldContainerProps> {
         <div className="ds-u-visibility--screen-reader">{errorMessage}</div>
       ) : null;
 
-    // Field input props handled by <FieldContainer>
+    // Field input props handled by <FormControl>
+    // TODO: Move `setRef` logic into <TextField> and <Select>
+    // TODO: Simplify `id` props by using a singular `fieldId` prop, and a consistent id naming convention
+    // TODO: Use React Context to provide shared form props like `errorPlacement`, `inversed`, `fieldId`
     const fieldInputProps = {
       id: this.id,
       labelId: this.labelId,
@@ -191,7 +193,7 @@ export class FieldContainer extends React.Component<FieldContainerProps> {
   }
 }
 
-export const FieldContainerPropKeys = [
+export const FormControlPropKeys = [
   'className',
   'component',
   'errorId',
@@ -211,4 +213,4 @@ export const FieldContainerPropKeys = [
   'render',
 ];
 
-export default FieldContainer;
+export default FormControl;
