+++ packages/design-system/src/components/TextField/TextField.jsx
@@ -1,4 +1,4 @@
-import { FieldContainer, FieldContainerPropKeys } from '../FieldContainer/FieldContainer';
+import { FormControl, FormControlPropKeys } from '../FormControl/FormControl';
 import { omit, pick } from 'lodash';
 import PropTypes from 'prop-types';
 import React from 'react';
@@ -22,8 +22,8 @@ export class TextField extends React.PureComponent {
   }
 
   render() {
-    const containerProps = pick(this.props, FieldContainerPropKeys);
-    const inputOnlyProps = omit(this.props, FieldContainerPropKeys);
+    const containerProps = pick(this.props, FormControlPropKeys);
+    const inputOnlyProps = omit(this.props, FormControlPropKeys);
 
     // Add clearfix class
     const containerClassName = classNames(
@@ -32,11 +32,11 @@ export class TextField extends React.PureComponent {
     );
 
     // Use errorPlacement feature flag for <TextInput>
-    // Duplicate of errorPlacement defaulting that occurs inside <FieldContainer>
+    // Duplicate of errorPlacement defaulting that occurs inside <FormControl>
     const errorPlacement = this.props.errorPlacement || errorPlacementDefault();
 
     return (
-      <FieldContainer
+      <FormControl
         {...containerProps}
         className={containerClassName}
         component="div"
